hypolab/requirements.txt
django
djangorestframework
djangorestframework-simplejwt

# DB
psycopg2-binary

# message broker
confluent-kafka

# test
pytest
pytest-django
django-cors-headers
